requirements.txt
fastapi==0.116.1
msal==1.33.0
python-dotenv==1.1.1
requests~=2.32.5
uvicorn[standard]